Do not reload vimrc on BundleInstall
BundleInstall sources the current vimrc in order to reload all Bundles
in order to consider new Bundles that might have been added by the user
directly on their vimrc file. This, however, should not be Vundle's
responsibility, changes in vimrc should not be applied unless it is
re-sourced explicitly by the user.

This makes Vundle use the known list of Bundles for installation.