added support for options to the Bundle command
Wrote body for s:parse_options to allow users to specify config options
to the Bundle command. The most common use for this would be to set the
path to a plugins runtime files directory. If a plugin repo stores it's
runtime files in a subdirectory, you need to be able to specify the path
to this directory in order to laod the plugin. For exmaple,

" exvim stores runtime files in subdir of repo root
Bundle 'jwu/exvim' 'rtpath=./vimfiles'

The parser will read arbitary options, but currently only rtpath is
used.