Update line highlight links for doc help sections
Locking these to a specific treeish will keep them stable. The links
will have to be updated to refer to current docs as they are updated,
but at least clicking a link won't point you to something that may not
pertain at all to what you were expecting. Plus it was obviously already
necessary to maintain updated links one way or another...