Limit default rtp rearrangement
Rearranging rtp so that **all** default directories appear first has the
undesired side effect of not allowing plugins to override default syntax
files.

This changeset limits the rearrangement of the rtp to simply making sure
the first directory in the default runtimepath appears first in also in
the runtimepath after Vundle has finished manipulating it.

This should keep the original bugs fixed and should eliminate the
aforementioned undesired side effect.