Don't flag name duplication when installing the same plugin twice
Name duplication is checked so that users don't end up overwriting one
plugin directory with the contents from another plugin that shares a
common name.

However, if the repeated plugin happens to be the same, there will be no
surprises and there is no need to flag this as an error. This needs if
particularly acute due to the fact that we currently don't show, in the
plugin search, which of the resulting plugins are already installed, so a
user might very well attempt to install a plugin they already have.

With this change, instead of seeing an error in this scenario, the user
will see a message indicating the plugin is already installed, and no
additional cloning will be performed.

Fixes #509