On Windows, Vim will sometimes have no idea where $HOME actually is. Since the default for Vim user scripts on Windows is "$VIM/vimfiles/", have Vundle direct bundles to that path instead.
On some Windows machines, $HOME is modified to direct to a shared drive or folder, and causes Vim to load to that location, as opposed to directing to the user's actual home directory. This is to avoid cloning Vim scripts to what could be shared drives for collaboration & file sharing.

Signed-off-by: Lee Savide <laughingman182@gmail.com>