Set GIT_ASKPASS environment variable before clone
Right now, if you has a mistyped github repo in your vimrc and you run
:PluginInstall, then vundle hangs and asks for your github username and
password. This is caused because git wants to check to make sure the
error isn't caused by some permissions problem.

In #662 and #667, it was noted that we could fix this by setting the
GIT_ASKPASS environment variable.

In this commit, I set this environment variable for the git clone
commands.

Note: resolves #667